split out some code bits
